disable performance-stats feature by default
