+++ utils/near-performance-metrics/src/lib.rs
@@ -1,3 +1,15 @@
-pub mod actix;
 pub mod process;
-pub mod stats;
+pub mod stats_disabled;
+pub mod stats_enabled;
+pub mod actix_disabled;
+pub mod actix_enabled;
+
+#[cfg(feature = "performance_stats")]
+pub use stats_enabled as stats;
+#[cfg(feature = "performance_stats")]
+pub use actix_enabled as actix;
+
+#[cfg(not(feature = "performance_stats"))]
+pub use stats_disabled as stats;
+#[cfg(not(feature = "performance_stats"))]
+pub use actix_disabled as actix;
